+++ app/criteria/__init__.py
@@ -1,0 +1,9 @@
+from fillers_number import FillersNumberCriterion
+from fillers_ratio import FillersRatioCriterion
+from len_text_on_slide import LenTextOnSlideCriterion
+from number_slides import NumberSlidesCriterion
+from number_word_on_slide import NumberWordOnSlideCriterion
+from speech_duration import SpeechDurationCriterion
+from speech_is_not_in_database import SpeechIsNotInDatabaseCriterion
+from speech_pace import SpeechPaceCriterion
+from strict_speech_duration import StrictSpeechDurationCriterion
