+++ app/criteria/fillers_number/__init__.py
@@ -1,0 +1,1 @@
+from .criterion import FillersNumberCriterion
