+++ app/criteria/fillers_ratio/__init__.py
@@ -1,0 +1,1 @@
+from .criterion import FillersRatioCriterion
