+++ app/criteria/len_text_on_slide/__init__.py
@@ -1,0 +1,1 @@
+from .critetion import LenTextOnSlideCriterion
