+++ app/criteria/number_slides/__init__.py
@@ -1,0 +1,1 @@
+from .critetion import NumberSlidesCriterion
