+++ app/criteria/number_word_on_slide/__init__.py
@@ -1,0 +1,1 @@
+from .critetion import NumberWordOnSlideCriterion
