+++ app/criteria/speech_duration/__init__.py
@@ -1,0 +1,1 @@
+from .criterion import SpeechDurationCriterion
