+++ app/criteria/speech_is_not_in_database/__init__.py
@@ -1,0 +1,1 @@
+from .criterion import SpeechIsNotInDatabaseCriterion
