+++ app/criteria/speech_pace/__init__.py
@@ -1,0 +1,1 @@
+from .criterion import SpeechPaceCriterion
